+++ packages/react-ui-kit/src/Form/Input.tsx
@@ -23,10 +23,11 @@ import {TextTransformProperty} from 'csstype';
 import React from 'react';
 import {COLOR} from '../Identity';
 import {Theme} from '../Layout';
-import {TextProps} from '../Text';
+import {TextProps, Label} from '../Text';
 import {filterProps} from '../util';
 
 export interface InputProps<T = HTMLInputElement> extends TextProps<T> {
+  label?: string;
   markInvalid?: boolean;
   placeholderTextTransform?: TextTransformProperty;
 }
@@ -52,7 +53,10 @@ export const inputStyle: <T>(theme: Theme, props: InputProps<T>) => ObjectInterp
     '&::-webkit-input-placeholder': {
       ...placeholderStyle,
     },
-    '&:invalid': !markInvalid
+    '&:focus': {
+      boxShadow: `0 0 0 1px ${COLOR.BLUE}`,
+    },
+    '&:invalid:not(:focus)': !markInvalid
       ? {
           boxShadow: 'none',
         }
@@ -73,18 +77,29 @@ export const inputStyle: <T>(theme: Theme, props: InputProps<T>) => ObjectInterp
   };
 };
 
-export const INPUT_CLASSNAME = 'input';
-const filterInputProps = (props: InputProps) => filterProps(props, ['markInvalid', 'placeholderTextTransform']);
+export const INPUT_CLASSNAME = 'wireinput';
+const filterInputProps = (props: InputProps) =>
+  filterProps(props, ['markInvalid', 'placeholderTextTransform', 'label']);
 
 export const Input: React.FC<InputProps<HTMLInputElement>> = React.forwardRef<
   HTMLInputElement,
   InputProps<HTMLInputElement>
->(({type, ...props}, ref) => (
-  <input
-    className={INPUT_CLASSNAME}
-    css={theme => inputStyle(theme, props)}
-    ref={ref}
-    type={type}
-    {...filterInputProps(props)}
-  />
-));
+>(({type, ...props}, ref) => {
+  const input = (
+    <input
+      className={INPUT_CLASSNAME}
+      css={theme => inputStyle(theme, props)}
+      ref={ref}
+      type={type}
+      {...filterInputProps(props)}
+    />
+  );
+  return props.label ? (
+    <Label bold={false} markInvalid={props.markInvalid}>
+      <span style={{margin: '16px'}}>{props.label}</span>
+      {input}
+    </Label>
+  ) : (
+    input
+  );
+});